chore: add template inheritance layer
The inheritance will be
base -> portfolio -> index/index-cs.

The added portfolio layer is to enable separation of language texts in en and cs, while preventing duplication of the layout of index page. portfolio.html will be the place to make layout changes, if the intended content stays the same, no more changes will be needed.