feat: handle project cards with bootstrap builtins, get texts with jinja
The collapsible is now implemented using Bootstrap Collapsible. Collapsing is set to be conditioned with class, instead of id, to avoid duplication of code (since the cards are generated in loop, adding collapsible onto ids would either create duplicate ids, or cause splitting of the loop into lists of projects that 1 are always shown 2 are in the collapsible).

Text is now passed through jinja variables.

Text of show button changes based on the current state of collapsible.

Add js block into base to allow insertion of current script.