feat: move language and color theme switchers to navbar on mobile too, move hamburger button to the right
The switchers were non-intuitively hidden in the collapsible menu on mobile. With these changes they are shown always, so users know they can change language or theme without looking into the collapsible navbar.

The hamburger button is now on the right, which might be easier for people to access even with one hand.